pyproject.toml
[tool.poetry]
name = "bookshelf"
version = "0.1.0"
description = ""
authors = ["Vladislav Atakhanov <eloh1mzee@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
python-dotenv = "^1.0.1"
werkzeug = "^3.0.3"
psycopg2 = "^2.9.9"
gunicorn = "^23.0.0"
google = "^3.0.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
openai = "^1.44.0"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.16"
rank-bm25 = "^0.2.2"
faiss-cpu = "^1.8.0.post1"
pandas = "^2.2.2"
flask-migrate = "^4.0.7"
flask-socketio = "^5.3.7"
google-cloud-vision = "^3.7.4"
flasgger = "^0.9.7.1"
flask-mail = "^0.10.0"
moviepy = "^1.0.3"
opencv-python = "^4.10.0.84"


[tool.poetry.scripts]
database = "app.models:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
